backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ── Database ──────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Auth ──────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# ── Anthropic / Claude ────────────────────────────────────────
anthropic==0.25.0

# ── Technical Indicators ──────────────────────────────────────
pandas
pandas-ta
numpy

# ── Email ─────────────────────────────────────────────────────
aiosmtplib==3.0.1

# ── HTTP client (broker APIs) ─────────────────────────────────
httpx==0.27.0
kiteconnect>=4.0.0
pyotp

# ── Config ────────────────────────────────────────────────────
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── Rate limiting ─────────────────────────────────────────────
slowapi==0.1.9

# ── Utilities ─────────────────────────────────────────────────
pytz==2024.1
python-dateutil==2.9.0
